Moved base manager from definitions to managers namespace and folder.
