+++ Assets/MixedRealityToolkit/_Core/Managers/BaseManager.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace Microsoft.MixedReality.Toolkit.Internal.Definitions
+namespace Microsoft.MixedReality.Toolkit.Internal.Managers
 {
     /// <summary>
     /// The base manager implements the IMixedRealityManager interface and provides default properties for all managers
