+++ Assets/MixedRealityToolkit/_Core/Managers/BaseManager.cs.meta
