+++ Assets/MixedRealityToolkit/_Core/Managers/MixedRealityEventManager.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Internal.Definitions;
 using Microsoft.MixedReality.Toolkit.Internal.Interfaces.Events;
 using Microsoft.MixedReality.Toolkit.Internal.Utilities.Async;
 using System.Collections.Generic;
